==> thd_challenge/00. data_investigation.ipynb <==
### Lets look at the data a bit

import pandas as pd
import numpy as np

df = pd.read_csv('ds_challenge_data.csv')

##### Some outputs are cleared for privacy

df

df.max()

df.min()

df.count()

df.isna().sum()

There are no nulls! maybe 0's mean null, e.g., prods with 0 weight

df[df['Transaction_ID']=='846-008']

one txn can have many products in it

df.describe()

df['Online_Order_Type'].value_counts()

df['Online_Order_Type_Desc'].value_counts()

these two columns need to be ignored

df['Transaction_Type'].value_counts()

df['Transaction_Sales_Type'].value_counts()

these two are good for doing one-hot encoding

df['SubClass_Desc'].value_counts()

this is a good to do some word-embedding

df['Product_Unit_Dimension'].value_counts()

df['Product_Volumn_Unit'].value_counts()

use these to convert the values into one thing, FT3 is square feet probably
