==> thd_challenge/01. data_cleaing.ipynb <==
## Data Cleaing

1. convert all production dimentions to metric (grams, meters, CM)
2. Ignore unwanted columns that have only one value, e.g., Online_Order_Type and Online_Order_Type_Desc have value “ONLINE”
3. break down data and time columns to add in week number, month number, year, ...
4. bring in holidays into the dataset (Black Friday, etc.)

import pandas as pd
import numpy as np
from datetime import datetime
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('ds_challenge_data_fixed.csv')

#output cleared
df.sort_values(by=['Hash_Trans_ID'], inplace=False)

df.columns

### Cleanup the transaction id

#output cleared
df[df['Hash_Trans_ID']=='0006d7dd4c3fd47287aca3bbeb1976cb4edc4a7d']

#no need for this anymore
#df['Cleaned_Transaction_ID'] = df[['Transaction_ID','Transaction_Date_ID','Transaction_Hour_ID']].apply(
#    lambda x: '-'.join(x.astype(str)),
#    axis=1
#)

test1 = df.groupby('Hash_Trans_ID').agg({'PRO': np.min}).reset_index().rename({'PRO': 'MIN_PRO'}, axis=1)
test2 = df.groupby('Hash_Trans_ID').agg({'PRO': np.max}).reset_index().rename({'PRO': 'MAX_PRO'}, axis=1)

test = pd.merge(test1, test2, how='inner', on = 'Hash_Trans_ID')

test[test['MIN_PRO']!=test['MAX_PRO']]

Now - with new HashID everything seems to be fine


Old - At least seems that PROs are unique for these IDs, but this might create an issue where both PROs are the same but it is actually same transaction happened in a short amount of time from each other

### Convert everything to metric

df.columns

#### Prod Weight

df['Product_Weight_Dimension'].value_counts()

df['Product_Weight_Grams'] = df[['Product_Weight','Product_Weight_Dimension']].apply(
    lambda x: x[0]*1000 if x[1]=='KG' else (x[0]*453.592 if x[1]=='LB' else (x[0]*28.35 if x[1]=='OZ' else x[0])),
    axis=1
)

#test to see if it worked - cleared
df[ (df['Product_Weight_Dimension']=='KG') & (df['Product_Weight']!=0)]

#### Prod Volume

df['Product_Volumn_Unit'].value_counts()

df['Product_Volumn_Metric'] = df[['Product_Volumn','Product_Volumn_Unit']].apply(
    lambda x: x[0]*10.764 if x[1]=='FT3' else  x[0],
    axis=1
)

#### Prod Unit Dims

df['Product_Unit_Dimension'].value_counts()

df['Product_Height_CM'] = df[['Product_Height','Product_Unit_Dimension']].apply(
    lambda x: x[0]*30.48 if x[1]=='FT' else (x[0]*2.54 if x[1]=='IN' else x[0]),
    axis=1
)

df['Product_Length_CM'] = df[['Product_Length','Product_Unit_Dimension']].apply(
    lambda x: x[0]*30.48 if x[1]=='FT' else (x[0]*2.54 if x[1]=='IN' else x[0]),
    axis=1
)

df['Product_Width_CM'] = df[['Product_Width','Product_Unit_Dimension']].apply(
    lambda x: x[0]*30.48 if x[1]=='FT' else (x[0]*2.54 if x[1]=='IN' else x[0]),
    axis=1
)

## Keep only the columns that are required so far

df.columns

df['Online_Order_Type'].value_counts()

df['Online_Order_Type_Desc'].value_counts()

df = df[['Hash_Trans_ID','SKU_ID', 'Transaction_Date_ID', 'Transaction_Hour_ID','Transaction_Type',\
        'Transaction_Sales_Type', 'POS_Sales', 'POS_UnitsSold', 'Department_ID',\
        'Class_ID', 'SubClass_ID', 'SubClass_Desc', 'Product_Weight_Grams', \
        'Product_Height_CM', 'Product_Length_CM','Product_Width_CM', \
         'Product_Volumn_Metric','PRO']]

df.describe()

df.dtypes

## Cleanup Date and Time Columns

df['Transaction_Year'] = df.Transaction_Date_ID.astype(str).str.slice(0,4).astype(int)

df['Transaction_Month'] = df.Transaction_Date_ID.astype(str).str.slice(4,6).astype(int)

df['Transaction_Day'] = df.Transaction_Date_ID.astype(str).str.slice(6,8).astype(int)

#### break down the time column

df['Transaction_Time'] = df['Transaction_Hour_ID'].astype(str).apply(
    lambda x: '{0:0>6}'.format(x)
)

df['Transaction_Hour'] = df.Transaction_Time.astype(str).str.slice(0,2).astype(int)

df['Transaction_Min'] = df.Transaction_Time.astype(str).str.slice(2,4).astype(int)

df['Transaction_Sec'] = df.Transaction_Time.astype(str).str.slice(4,6).astype(int)

df.describe()

df['Transaction_Date']=df[['Transaction_Year','Transaction_Month','Transaction_Day','Transaction_Hour','Transaction_Min','Transaction_Sec']].apply(
    lambda x: datetime(x[0],x[1],x[2],x[3],x[4],x[5]),
    axis=1
)

df['Transaction_Date'] = pd.to_datetime(df['Transaction_Date'])

df['Transaction_Day_of_Week'] = df['Transaction_Date'].dt.dayofweek

df['Transaction_Week'] = df['Transaction_Date'].dt.weekofyear

df['Transaction_Day_of_Year'] = df['Transaction_Date'].dt.dayofyear

### Add in Holidays to the dataset

from datetime import date 
import holidays

hldy = holidays.Canada()

df['holiday_nm'] = df['Transaction_Date'].apply(lambda x: hldy.get(x))

df.columns

df = df[['Hash_Trans_ID','SKU_ID', 'Transaction_Date', 'Transaction_Year', 'Transaction_Month', 'Transaction_Day',\
         'Transaction_Hour','Transaction_Min', 'Transaction_Sec','Transaction_Day_of_Week',\
       'Transaction_Week', 'Transaction_Day_of_Year','holiday_nm','Transaction_Type',\
        'Transaction_Sales_Type', 'POS_Sales', 'POS_UnitsSold', 'Department_ID',\
        'Class_ID', 'SubClass_ID', 'SubClass_Desc', 'Product_Weight_Grams', \
        'Product_Height_CM', 'Product_Length_CM','Product_Width_CM', \
         'Product_Volumn_Metric','PRO']]

df.to_csv('ds_challenge_data_cleaned.csv', index=False)
