==> thd_challenge/02. feature_engineering.ipynb <==
# Feature Engineering

1. create price feature based on POS sales and POS dollars
2. Impute nulls in the dataset
3. Vectorization
4. Moving Average Feature

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.cluster import KMeans
from sklearn.metrics import adjusted_rand_score
import warnings

warnings.filterwarnings('ignore')

df = pd.read_csv('ds_challenge_data_cleaned.csv')

df.columns

## Pricing features

df['Price'] = df['POS_Sales']/df['POS_UnitsSold']

#### create SKU level price features seperate for each year

#Main assumption is that we know the price one year before
sku_df = df.groupby(['SKU_ID','Transaction_Year']).agg({'Price':np.mean}).reset_index()
sku_df['SKU_AVG_Price'] = sku_df['Price']
sku_df = sku_df[['SKU_ID','Transaction_Year','SKU_AVG_Price']]

sku_df2 = df.groupby(['SKU_ID','Transaction_Year']).agg({'Price':np.min}).reset_index()
sku_df2['SKU_MIN_Price'] = sku_df2['Price']
sku_df2 = sku_df2[['SKU_ID','Transaction_Year','SKU_MIN_Price']]

sku_df3 = df.groupby(['SKU_ID','Transaction_Year']).agg({'Price':np.max}).reset_index()
sku_df3['SKU_MAX_Price'] = sku_df3['Price']
sku_df3 = sku_df3[['SKU_ID','Transaction_Year','SKU_MAX_Price']]

sku_df = sku_df.merge(sku_df2, on=['SKU_ID','Transaction_Year']).merge(sku_df3, on=['SKU_ID','Transaction_Year'])
df = df.merge(sku_df, on=['SKU_ID','Transaction_Year'], how='left')

# Vectorization

Data is better to be aggregated to transaction level instead of transaction-sku level
At transaction level vectors that are built based on product attributes are immune to problems like similar SKUs and product cannibalization 


#Remove Nulls to start Vectorization
#df.fillna('none', inplace = True)
df.fillna(0, inplace = True)

### Txn Type, Sales Type and Holiday

df.groupby('Transaction_Type')['PRO'].mean()

df[((df['Transaction_Type']=='Pro Desk - Refund') & (df['PRO']==0))]

df.groupby('holiday_nm')['PRO'].mean()

df.groupby('Transaction_Sales_Type')['PRO'].mean()

df.groupby('Department_ID')['PRO'].mean()

### Clustering SubClass Desc

df['SubClass_Desc_Clean'] = df['SubClass_Desc'].apply(lambda x: ''.join([i for i in x if not i.isdigit()]))

df['SubClass_Desc_Clean']= df['SubClass_Desc_Clean'].apply(lambda x: x.lower().replace('XX','').replace('xx','')  )

df['SubClass_Desc_Clean']= df['SubClass_Desc_Clean']\
        .apply(lambda x: x.replace('&','').replace('-','').replace('/','').replace('$','')\
                .replace('%','').replace(' ','').replace("'",'').replace(".",'').strip()  )

#vectorizer = TfidfVectorizer(stop_words='english')
vectorizer = CountVectorizer(stop_words="english", analyzer='char', ngram_range=(3, 5), max_df=1.0, min_df=1, max_features=None)
X = vectorizer.fit_transform(df['SubClass_Desc_Clean'])

true_k = 10
model = KMeans(n_clusters=true_k, init='k-means++', max_iter=100, n_init=1, n_jobs=-1)
model.fit(X)

df['SubClass_Desc_Cluster'] = df['SubClass_Desc_Clean'].apply(lambda x: model.predict(vectorizer.transform([x])).astype('int')[0] )

df.groupby(['SubClass_Desc_Cluster']).count()['Hash_Trans_ID']

df.groupby(['SubClass_Desc_Cluster'])['PRO'].mean()

df = pd.get_dummies(data=df, columns = ['Transaction_Type','Transaction_Sales_Type','Department_ID','SubClass_Desc_Cluster'])

df.columns

## Put back everything into Transaction level

df_final = df.groupby('Hash_Trans_ID').agg({'SKU_ID':np.size,
                                            'Transaction_Date':np.max,
                                            'Transaction_Year':np.max, 'Transaction_Month':np.max,
                                            'Transaction_Day':np.max, 'Transaction_Hour':np.max, 
                                            'Transaction_Day_of_Week':np.max,
                                            'Transaction_Week':np.max, 
                                            'Transaction_Day_of_Year':np.max, 
                                            'holiday_nm':np.max,
                                            'PRO':np.max}).reset_index()

df_final = df_final.rename(columns={'SKU_ID': 'SKU_ID_Count'})

### Create Sum/Min/Max/Std Vectors 

df_sum = df.groupby('Hash_Trans_ID').agg({
                                'POS_Sales':np.sum,
                                'POS_UnitsSold':np.sum, 
                                'Product_Length_CM':np.sum,
                                'Product_Height_CM':np.sum,
                                'Product_Width_CM':np.sum,
                                'Product_Weight_Grams':np.sum, 
                                'Product_Volumn_Metric':np.sum,
                                'Price':np.sum, 
                                'SKU_AVG_Price':np.sum,
                                'SKU_MIN_Price':np.sum, 
                                'SKU_MAX_Price':np.sum, 
                                'Transaction_Type_1':np.sum,
                                'Transaction_Type_2':np.sum,
                                'Transaction_Type_3':np.sum,
                                'Transaction_Type_4':np.sum,
                                'Transaction_Type_5':np.sum,
                                'Transaction_Type_6':np.sum,
                                'Transaction_Type_7':np.sum,
                                'Transaction_Type_8':np.sum,
                                'Transaction_Type_9':np.sum,
                                'Transaction_Type_10':np.sum,
                                'Transaction_Type_11':np.sum,
                                'Transaction_Sales_Type_12':np.sum,
                                'Transaction_Sales_Type_13':np.sum, 
                                'Transaction_Sales_Type_14':np.sum,
                                'Transaction_Sales_Type_15':np.sum,
                                'Transaction_Sales_Type_16':np.sum,
                                'Department_ID_13':np.sum, 'Department_ID_16':np.sum, 'Department_ID_17':np.sum,
                                'Department_ID_21':np.sum, 'Department_ID_22':np.sum, 'Department_ID_23':np.sum,
                                'Department_ID_24':np.sum, 'Department_ID_25':np.sum, 'Department_ID_26':np.sum,
                                'Department_ID_27':np.sum, 'Department_ID_28':np.sum, 'Department_ID_29':np.sum,
                                'Department_ID_30':np.sum, 'Department_ID_59':np.sum, 'Department_ID_78':np.sum,
                                'SubClass_Desc_Cluster_0':np.sum, 'SubClass_Desc_Cluster_1':np.sum,
                                'SubClass_Desc_Cluster_2':np.sum, 'SubClass_Desc_Cluster_3':np.sum,
                                'SubClass_Desc_Cluster_4':np.sum, 'SubClass_Desc_Cluster_5':np.sum,
                                'SubClass_Desc_Cluster_6':np.sum, 'SubClass_Desc_Cluster_7':np.sum,
                                'SubClass_Desc_Cluster_8':np.sum, 'SubClass_Desc_Cluster_9':np.sum
                                }).reset_index()

df_sum = df_sum.rename(columns={'POS_Sales': 'POS_Sales_Sum',
                                    'POS_UnitsSold': 'POS_UnitsSold_Sum', 'Product_Length_CM': 'Product_Lengthe_Sum',
                                    'Product_Width_CM': 'Product_Widthe_Sum', 'Product_Weight_Grams': 'Product_Weighte_Sum',
                                    'Price': 'Price_Sum',  
                                    'SKU_AVG_Price':'SKU_AVG_Price_Sum',
                                    'SKU_MIN_Price':'SKU_MIN_Price_Sum', 
                                    'SKU_MAX_Price':'SKU_MAX_Price_Sum',
                                    'Product_Height_CM': 'Product_Heighte_Sum',
                                    'Product_Volumn_Metric': 'Product_Volumee_Sum', 'Product_Height_CM': 'Product_Heighte_Sum'})

df_final = df_final.merge(df_sum, on=['Hash_Trans_ID'], how='left')

df_mean = df.groupby('Hash_Trans_ID').agg({
                                'Price':np.mean, 
                                'SKU_AVG_Price':np.mean,
                                'SKU_MIN_Price':np.mean, 
                                'SKU_MAX_Price':np.mean, 
                                'POS_Sales':np.mean,
                                'POS_UnitsSold':np.mean, 
                                'Product_Length_CM':np.mean, 
                                'Product_Width_CM':np.mean,
                                'Product_Weight_Grams':np.mean, 
                                'Product_Height_CM':np.mean,
                                'Product_Volumn_Metric':np.mean}).reset_index()

df_mean = df_mean.rename(columns={'Price':'Price_Avg', 
                                'SKU_AVG_Price':'SKU_AVG_Price_Avg',
                                'SKU_MIN_Price':'SKU_MIN_Price_Avg', 
                                'SKU_MAX_Price':'SKU_MAX_Price_Avg', 
                                'POS_Sales':'POS_Sales_Avg',
                                'POS_UnitsSold':'POS_UnitsSold_Avg', 
                                'Product_Length_CM':'Product_Length_Avg', 
                                'Product_Width_CM':'Product_Width_Avg',
                                'Product_Weight_Grams':'Product_Weight_Avg', 
                                'Product_Height_CM':'Product_Height_Avg',
                                'Product_Volumn_Metric':'Product_Volumn_Avg'})

df_final = df_final.merge(df_mean, on=['Hash_Trans_ID'], how='left')

df_max = df.groupby('Hash_Trans_ID').agg({
                                'Price':np.max, 
                                'SKU_AVG_Price':np.max,
                                'SKU_MIN_Price':np.max, 
                                'SKU_MAX_Price':np.max,
                                'POS_Sales':np.max,'POS_UnitsSold':np.max, 
                                'Product_Length_CM':np.max, 'Product_Width_CM':np.max,
                                'Product_Weight_Grams':np.max, 'Product_Height_CM':np.max,
                                'Product_Volumn_Metric':np.max}).reset_index()

df_max = df_max.rename(columns={'POS_Sales': 'POS_Sales_Max', 'POS_UnitsSold': 'POS_UnitsSold_Max',
                                    'Product_Length_CM': 'Product_Length_Max', 'Product_Width_CM': 'Product_Width_Max',
                                    'Product_Weight_Grams': 'Product_Weight_Max', 
                                    'Price': 'Price_Max',  
                                    'SKU_AVG_Price':'SKU_AVG_Price_Max',
                                    'SKU_MIN_Price':'SKU_MIN_Price_Max', 
                                    'SKU_MAX_Price':'SKU_MAX_Price_Max',
                                    'Product_Height_CM': 'Product_Height_Max', 'Product_Volumn_Metric': 'Product_Volume_Max'})

df_final = df_final.merge(df_max, on=['Hash_Trans_ID'], how='left')

df_min = df.groupby('Hash_Trans_ID').agg({
                                'Price':np.min, 
                                'SKU_AVG_Price':np.min,
                                'SKU_MIN_Price':np.min, 
                                'SKU_MAX_Price':np.min,
                                'POS_Sales':np.min,'POS_UnitsSold':np.min, 
                                'Product_Length_CM':np.min, 'Product_Width_CM':np.min,
                                'Product_Weight_Grams':np.min, 'Product_Height_CM':np.min,
                                'Product_Volumn_Metric':np.min}).reset_index()

df_min = df_min.rename(columns={'POS_Sales': 'POS_Sales_Min', 'POS_UnitsSold': 'POS_UnitsSold_Min',
                                    'Product_Length_CM': 'Product_Length_Min', 'Product_Width_CM': 'Product_Width_Min',
                                    'Product_Weight_Grams': 'Product_Weight_Min', 
                                    'Price': 'Price_Min',  
                                    'SKU_AVG_Price':'SKU_AVG_Price_Min',
                                    'SKU_MIN_Price':'SKU_MIN_Price_Min', 
                                    'SKU_MAX_Price':'SKU_MAX_Price_Min',
                                    'Product_Height_CM': 'Product_Height_Min', 'Product_Volumn_Metric': 'Product_Volume_Min'})

df_final = df_final.merge(df_min, on=['Hash_Trans_ID'], how='left')

df_std = df.groupby('Hash_Trans_ID').agg({
                                'Price':np.std, 
                                'SKU_AVG_Price':np.std,
                                'SKU_MIN_Price':np.std, 
                                'SKU_MAX_Price':np.std,
                                'POS_Sales':np.std,'POS_UnitsSold':np.std, 
                                'Product_Length_CM':np.std, 'Product_Width_CM':np.std,
                                'Product_Weight_Grams':np.std, 'Product_Height_CM':np.std,
                                'Product_Volumn_Metric':np.std}).reset_index()

df_std = df_std.rename(columns={'POS_Sales': 'POS_Sales_Std', 'POS_UnitsSold': 'POS_UnitsSold_Std',
                                    'Product_Length_CM': 'Product_Length_Std', 'Product_Width_CM': 'Product_Width_Std',
                                    'Product_Weight_Grams': 'Product_Weight_Std', 
                                    'Price': 'Price_Std',  
                                    'SKU_AVG_Price':'SKU_AVG_Price_Std',
                                    'SKU_MIN_Price':'SKU_MIN_Price_Std', 
                                    'SKU_MAX_Price':'SKU_MAX_Price_Std',
                                    'Product_Height_CM': 'Product_Height_Std', 'Product_Volumn_Metric': 'Product_Volume_Std'})

df_final = df_final.merge(df_std, on=['Hash_Trans_ID'], how='left')

## ClassID and SubClassID Vectorization

df['basket_row'] = df.sort_values(['SKU_ID'], ascending=[True]) \
             .groupby(['Hash_Trans_ID']) \
             .cumcount() + 1
df['basket_row'] = df['basket_row'].apply(lambda x: 11 if x>10 else x)

pt = pd.pivot_table(df, values=['POS_Sales','Class_ID','SubClass_ID'], index='Hash_Trans_ID', 
                    columns=['basket_row'], aggfunc=np.max).reset_index()
pt.fillna(0, inplace = True)

df_final = df_final.merge(pt, on=['Hash_Trans_ID'], how='left')

# Moving Average Feature

df_final = df_final.sort_values(by=['Transaction_Date','Hash_Trans_ID'])

for col in ['Transaction_Type_Commercial Line Sale','Transaction_Type_Front line - Sales','Transaction_Type_Front line -Self Chk',
            'Transaction_Type_Garden Sales','Transaction_Type_Pro Desk - Refund','Transaction_Type_Pro Desk - Sales',
            'Transaction_Type_Returns desk - Refnd','Transaction_Type_Sp. Srvc. Dsk  VPOS','Transaction_Type_Sp. Srvc. Dsk Refnd',
            'Transaction_Type_Tool Rental -Refund','Transaction_Type_Tools - Refund']:
    df_cat = df_final[df_final[col]>0]
    df_cat = df_cat[['Transaction_Date','PRO']].rolling(50, on='Transaction_Date').mean()
    df_cat.fillna(0, inplace = True)
    df_cat = df_cat.groupby('Transaction_Date').mean().reset_index()
    df_cat = df_cat.rename(columns={'PRO':'Moving_Avg_Txn_'+col})
    df_final = df_final.merge(df_cat, on=['Transaction_Date'], how='left')

df_final.fillna(0, inplace = True)

## Write the outputs for modeling

df_final.to_csv('ds_challenge_data_final.csv', index=False, sep='\t')
