==> thd_challenge/03.2 Hyper_Parameter_Optimization.ipynb <==
#Import libraries:
import pandas as pd
import numpy as np
import xgboost as xgb
from xgboost.sklearn import XGBClassifier
from sklearn.model_selection import cross_validate
from sklearn import metrics
import matplotlib.pylab as plt
%matplotlib inline
from matplotlib.pylab import rcParams
rcParams['figure.figsize'] = 12, 4

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('ds_challenge_data_final.csv',sep='\t')
#df.fillna(0, inplace = True)
#df.fillna("none", inplace = True)
#df = df.sample(frac=0.05)


df['holiday_nm'] = df['holiday_nm'].apply(lambda x: 'none' if ((str(x).strip()=='0') | (x==0)) else x)

target = 'PRO'
cols_to_drop = ['Transaction_Date','Hash_Trans_ID','PRO' ]
cat_cols = ['holiday_nm']
df = pd.get_dummies(data=df, columns = cat_cols)

df_train = df[df['Transaction_Year'] < 2015]
train_targets = df_train['PRO'].as_matrix().astype('float32')
train_data = df_train.drop(cols_to_drop, axis=1, inplace=False).as_matrix().astype('float32')
df_test = df[df['Transaction_Year'] == 2015]
test_targets = df_test['PRO'].as_matrix().astype('float32')
test_data = df_test.drop(cols_to_drop, axis=1, inplace=False).as_matrix().astype('float32')

### Crossvalidation training/testing

def modelfit(alg, dtrain, dtest, predictors,useTrainCV=True, cv_folds=5, early_stopping_rounds=50):
    
    if useTrainCV:
        xgb_param = alg.get_xgb_params()
        xgtrain = xgb.DMatrix(dtrain[predictors].values, label=dtrain[target].values)
        cvresult = xgb.cv(xgb_param, xgtrain, num_boost_round=alg.get_params()['n_estimators'], nfold=cv_folds,
            metrics='auc', early_stopping_rounds=early_stopping_rounds)
        alg.set_params(n_estimators=cvresult.shape[0])
    
    #Fit the algorithm on the data
    alg.fit(dtrain[predictors], dtrain[target],eval_metric='auc')
        
    #Predict training set:
    dtrain_predictions = alg.predict(dtrain[predictors])
    dtrain_predprob = alg.predict_proba(dtrain[predictors])[:,1]
    dtest_predictions = alg.predict(dtest[predictors])
    dtest_predprob = alg.predict_proba(dtest[predictors])[:,1]
        
    #Print model report:
    print("Accuracy  (Train): %.4g" % metrics.accuracy_score(dtrain[target].values, dtrain_predictions))
    print("AUC Score (Train): %f" % metrics.roc_auc_score(dtrain[target], dtrain_predprob))
    print("Accuracy  (Test): %.4g" % metrics.accuracy_score(dtest[target].values, dtest_predictions))
    print("AUC Score (Test): %f" % metrics.roc_auc_score(dtest[target], dtest_predprob))

### get the baseline

predictors = [x for x in df_train.columns if x not in cols_to_drop]

xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, learning_rate=0.2, n_estimators=100, verbosity=1, n_jobs=-1)
modelfit(xgb1, df_train, df_test, predictors)

## Tune Max_Depth and Min_Child_Weight

for max_depth in [2,3,4,5,6,7,8]:
    for min_child_weight in [1,2,3,4,5]:
        print("Max Depth: "+str(max_depth)+" Min Child Weight: "+str(min_child_weight))
        xgb1 = xgb.XGBClassifier(max_depth=max_depth, min_child_weight=min_child_weight, learning_rate=0.2, n_estimators=100, verbosity=1, n_jobs=-1)
        modelfit(xgb1, df_train, df_test, predictors)

## Tune Gamma

for gamma in [0,1,2,5]:
    print("Gamma: "+str(gamma))
    xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, gamma=gamma, learning_rate=0.2, n_estimators=100, verbosity=1, n_jobs=-1)
    modelfit(xgb1, df_train, df_test, predictors)

## Tune subsample and colsample_bytree

for subsample in [0.75,0.90]:
    for colsample_bytree in [0.75,0.90]:
        print("subsample: "+str(subsample)+" colsample_bytree: "+str(colsample_bytree))
        xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, gamma=0, subsample=subsample, colsample_bytree=colsample_bytree,  learning_rate=0.2, n_estimators=100, verbosity=1, n_jobs=-1)
        modelfit(xgb1, df_train, df_test, predictors)

## Tuning Regularization Parameter (Alpha)

for reg_alpha in [1e-5, 1e-2, 0.1, 1, 100]:
    print("reg_alpha: "+str(reg_alpha))
    xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, reg_alpha=reg_alpha, gamma=0, subsample=0.9, colsample_bytree=0.75, learning_rate=0.2, n_estimators=100, verbosity=1, n_jobs=-1)
    modelfit(xgb1, df_train, df_test, predictors)

## Tune Learning Rate

for learning_rate in [0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0]:
    print("learning_rate: "+str(learning_rate))
    xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, reg_alpha=1e-5, gamma=0, subsample=0.9, colsample_bytree=0.75, learning_rate=learning_rate, n_estimators=100, verbosity=1, n_jobs=-1)
    modelfit(xgb1, df_train, df_test, predictors)

for learning_rate in [0.12,0.14,0.16,0.18,0.2,0.22]:
    print("learning_rate: "+str(learning_rate))
    xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, reg_alpha=1e-5, gamma=0, subsample=0.9, colsample_bytree=0.75, learning_rate=learning_rate, n_estimators=100, verbosity=1, n_jobs=-1)
    modelfit(xgb1, df_train, df_test, predictors)

## Tune Number of Estimators

for n_estimators in [5,10,25,50,100,250,500]:
    print("n_estimators: "+str(n_estimators))
    xgb1 = xgb.XGBClassifier(max_depth=2, min_child_weight=1, reg_alpha=1e-5, gamma=0, subsample=0.9, colsample_bytree=0.75, learning_rate=0.2, n_estimators=n_estimators, verbosity=1, n_jobs=-1)
    modelfit(xgb1, df_train, df_test, predictors)

## Best Model So Far

best_mdl = xgb.XGBClassifier(max_depth=2, min_child_weight=1, reg_alpha=1e-5, gamma=0,subsample=0.9, colsample_bytree=0.75, learning_rate=0.2, n_estimators=250, verbosity=1, n_jobs=-1)

modelfit(best_mdl, df_train, df_test, predictors)
